requirements.txt
pandas
pyspark
apache-airflow
psycopg2-binary
boto3